+++ scripts/install.ps1
@@ -1,3 +1,3 @@
 #!/usr/bin/env pwsh
 
-deno install --unstable -A -f -n scaffold https://deno.land/x/scaffold/cli.ts
+deno install --unstable -A -f -n scaffold https://deno.land/x/scaffold/scaffold.ts